EPR-20: creacion y configuracion de businesstype

+++ demo/src/main/java/com/example/demo/repository/BusinessTypeRepository.java
@@ -1,4 +1,4 @@
 package com.example.demo.repository;
 import com.example.demo.model.BusinessType;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface BusinessTypeRepository extends JpaRepository<BusinessType, Integer> { }
+public interface BusinessTypeRepository extends JpaRepository<BusinessType, Long> { }
